feat: implement permission middleware and apply it.

+++ src/routes/v1/dashboard/admin.route.ts
@@ -1,6 +1,7 @@
 import { Router } from "express";
 import AdminController from "../../../controllers/admin.controller";
 import { controllerAsync } from "../../../middlewares/handlers/controllerAsync.handler";
+import { checkPermission } from "../../../middlewares/permission.middleware";
 import AdminRepository from "../../../repositories/admin.repository";
 import AdminService from "../../../services/admin.service";
 
@@ -12,11 +13,13 @@ const adminController = new AdminController(adminService);
 router
 	.route("/")
 	.get([
+		checkPermission("admin", "read"),
 		controllerAsync((request, response) =>
 			adminController.getAllAdmin(request, response),
 		),
 	])
 	.post([
+		checkPermission("admin", "create"),
 		controllerAsync((request, response) =>
 			adminController.createAdmin(request, response),
 		),
@@ -25,16 +28,19 @@ router
 router
 	.route("/:id")
 	.get([
+		checkPermission("admin", "read"),
 		controllerAsync((request, response) =>
 			adminController.getSingleAdmin(request, response),
 		),
 	])
 	.put([
+		checkPermission("admin", "update"),
 		controllerAsync((request, response) =>
 			adminController.updateAdmin(request, response),
 		),
 	])
 	.delete([
+		checkPermission("admin", "delete"),
 		controllerAsync((request, response) =>
 			adminController.deleteAdmin(request, response),
 		),
